Extract code into components
