+++ public/app/plugins/datasource/influxdb/components/editor/config-v2/LeftSideBar.tsx
@@ -1,4 +1,4 @@
-import { Box, InlineField, LinkButton, Space, Stack, Text } from '@grafana/ui';
+import { Box, InlineField, LinkButton, PanelContainer, Space, Stack, Text } from '@grafana/ui';
 
 import { CONFIG_SECTION_HEADERS } from './constants';
 
@@ -34,7 +34,6 @@ export const LeftSideBar = () => {
     <Stack>
       <Box flex={1} marginY={5}>
         <Text element="h4">InfluxDB</Text>
-        <Space v={3} />
         {generateSectionHeaders()}
       </Box>
     </Stack>
